requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
django-filter==23.5
Pillow==10.2.0
django-allauth==0.60.1

